% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIE.R
\name{CIE}
\alias{CIE}
\title{Assess confounding effects by Change-in-Estimate approach}
\usage{
CIE(
  data,
  outcome = NULL,
  time = NULL,
  exposure = NULL,
  covariates = NULL,
  positive = "auto",
  model = c("auto", "linear", "logit", "cox", "poson", "logbinom"),
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{outcome}{outcome variable name.}

\item{time}{time variable name, for Cox regression.}

\item{exposure}{exposure variable name.}

\item{covariates}{covariate names, a vector or a list.}

\item{positive}{in which positive of outcome variable to make the comparison.
By default, positive is automatically defined. If outcome is a factor variable,
then positive is defined as the highest level. If outcome is a numerical
variable, then positive is defined as the largest value.}

\item{...}{further arguments.}
}
\value{
a data frame.
}
\description{
Assess confounding effects by Change-in-Estimate approach
}
