% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.R
\name{cox}
\alias{cox}
\title{Fitting Cox models}
\usage{
cox(data, formula, positive = "auto", ...)
}
\arguments{
\item{data}{a data.frame in which to interpret the variables named in the
formula, or in the subset and the weights argument.}

\item{formula}{a formula object, with the response on the left of a ~ operator,
and the terms on the right. The response must be a survival object as returned
by the Surv function. For a multi-state model the formula may be a list of formulas.}

\item{positive}{in which positive of outcome variable to make the comparison.
By default, positive is automatically defined. If outcome is a factor variable,
then positive is defined as the highest level. If outcome is a numerical
variable, then positive is defined as the largest value.}

\item{...}{Other arguments will be passed to \code{\link[survival:coxph]{survival::coxph()}}.}
}
\value{
an object of class coxph representing the fit.
}
\description{
Fits a Cox proportional hazards regression model. Time dependent variables,
time dependent strata, multiple events per subject, and other extensions are
incorporated using the counting process formulation of Andersen and Gill.
}
