% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.R
\name{cox2}
\alias{cox2}
\title{Fitting Cox models}
\usage{
cox2(
  data,
  outcome = NULL,
  time = NULL,
  exposure = NULL,
  covariates = NULL,
  positive = "auto",
  ...
)
}
\arguments{
\item{data}{a data frame, list or environment (or object coercible by as.data.
frame to a data frame) containing the variables in the model.}

\item{outcome}{outcome variable name.}

\item{time}{time variable name.}

\item{exposure}{exposure variable name.}

\item{covariates}{covariate names, a vector or a list.}

\item{positive}{in which positive of outcome variable to make the comparison.
By default, positive is automatically defined. If outcome is a factor variable,
then positive is defined as the highest level. If outcome is a numerical
variable, then positive is defined as the largest value.}

\item{...}{Other arguments will be passed to \code{\link[survival:coxph]{survival::coxph()}}.}
}
\value{
an object of class coxph representing the fit.
}
\description{
Fitting Cox models
}
