% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{gee_change}
\alias{gee_change}
\title{Change from baseline by GEE}
\usage{
gee_change(
  data,
  outcome = NULL,
  exposure = NULL,
  covariates = NULL,
  id = NULL,
  ci.change = TRUE,
  family = stats::gaussian(),
  corstr = "unstructured",
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{outcome}{outcome variable name.}

\item{exposure}{exposure variable name.}

\item{covariates}{covariate names, a vector or a list.}

\item{id}{which identifies the clusters.}

\item{ci.change}{ci for change.}

\item{family}{see corresponding documentation to \code{\link[=glm]{glm()}}}

\item{corstr}{a character string specifying the correlation structure.
The following are permitted: independence, exchangeable, ar1, unstructured,
and userdefined.}

\item{...}{further arguments passed to}
}
\value{
a data frame.
}
\description{
Change from baseline by GEE
}
