% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km.R
\name{km}
\alias{km}
\title{Kaplan-Meier method}
\usage{
km(
  data,
  outcome = NULL,
  time = NULL,
  exposure = NULL,
  covariates = NULL,
  positive = "auto",
  test = TRUE,
  median = TRUE,
  at = NULL,
  digits.median = 1,
  digits.pvalue = 3,
  overall = FALSE,
  ci = TRUE,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{outcome}{outcome variable name.}

\item{time}{time variable name, for Cox regression.}

\item{exposure}{exposure variable name.}

\item{covariates}{covariate names, a vector or a list.}

\item{positive}{in which positive of outcome variable to make the comparison.
By default, positive is automatically defined. If outcome is a factor variable,
then positive is defined as the highest level. If outcome is a numerical
variable, then positive is defined as the largest value.}

\item{test}{show log-rank test?}

\item{median}{show median survival?}

\item{at}{survival rate at points of time.}

\item{digits.median}{digits for median survival.}

\item{digits.pvalue}{digits for P value.}

\item{overall}{show the result of all patients?}

\item{ci}{show CI?}

\item{...}{unused.}
}
\value{
a data frame.
}
\description{
Kaplan-Meier method
}
