% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{linear}
\alias{linear}
\title{Fitting linear models}
\usage{
linear(data, formula, ...)
}
\arguments{
\item{data}{a data frame, list or environment (or object coercible by as.data.
frame to a data frame) containing the variables in the model.}

\item{formula}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted.}

\item{...}{additional arguments to be passed to \code{\link[=lm]{lm()}} function.}
}
\value{
an object of class "lm" or for multivariate ('multiple') responses
of class c("mlm", "lm").
}
\description{
linear is used to fit linear models, including multivariate ones. It can be
used to carry out regression, single stratum analysis of variance and analysis
of covariance (although aov may provide a more convenient interface for these).
}
