% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{linear2}
\alias{linear2}
\title{Fitting linear models}
\usage{
linear2(data, outcome = NULL, exposure = NULL, covariates = NULL, ...)
}
\arguments{
\item{data}{a data frame, list or environment (or object coercible by as.data.
frame to a data frame) containing the variables in the model.}

\item{outcome}{outcome variable name.}

\item{exposure}{exposure variable name.}

\item{covariates}{covariate names, a vector or a list.}

\item{...}{additional arguments to be passed to \code{\link[=lm]{lm()}} function.}
}
\value{
an object of class "lm" or for multivariate ('multiple') responses
of class c("mlm", "lm").
}
\description{
Fitting linear models
}
