% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Fitting logistic models}
\usage{
logit(data, formula, positive = "auto", ...)
}
\arguments{
\item{data}{a data frame, list or environment (or object coercible by
as.data.frame to a data frame) containing the variables in the model.}

\item{formula}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted.}

\item{positive}{in which positive of outcome variable to make the comparison.
By default, positive is automatically defined. If outcome is a factor variable,
then positive is defined as the highest level. If outcome is a numerical
variable, then positive is defined as the largest value.}

\item{...}{additional arguments to be passed to \code{\link[=glm]{glm()}} function.}
}
\value{
an object of class inheriting from "glm" which inherits from the
class "lm".
}
\description{
Fitting logistic models
}
