% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom.R
\name{multinom}
\alias{multinom}
\title{Fitting multinomial logistic models}
\usage{
multinom(data, formula, ...)
}
\arguments{
\item{data}{a data frame, list or environment (or object coercible by
as.data.frame to a data frame) containing the variables in the model.}

\item{formula}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted.}

\item{...}{additional arguments to be passed to \code{\link[=glm]{glm()}} function.}
}
\description{
Fitting multinomial logistic models
}
\seealso{
\code{\link[nnet:multinom]{nnet::multinom()}}
}
