% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariable.R
\name{multivariable.coxph}
\alias{multivariable.coxph}
\title{Multivariable analysis for coxph}
\usage{
\method{multivariable}{coxph}(
  data,
  outcome = NULL,
  time = NULL,
  indepts = NULL,
  model = c("cox"),
  effect.values = c("b", "se", "effect", "p"),
  conf.level = 0.95,
  conf.brackets = NULL,
  conf.separator = NULL,
  digits.pvalue = 3,
  digits.effect = 2,
  ref.value = "Referrence",
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{outcome}{outcome variable name.}

\item{time}{time variable name, for Cox regression.}

\item{indepts}{independent variable names.}

\item{model}{regression models.}

\item{effect.values}{Effect value, 'ne' for No. of event, 'nt' for No. of
total, 'net' for No. of event and total, 'nne' for No. of non-event, 'wald'
for Wald Chi-square value, 'B' for coefficients, 'p' for P value, 'se' for
standard error, 'or' for odds ratio and its 95\% CI, ignore case.}

\item{conf.level}{The confidence level to use for the confidence interval if
conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults to
0.95, which corresponds to a 95 percent confidence interval.}

\item{conf.brackets}{brackets of CI format.}

\item{conf.separator}{separate of CI format.}

\item{digits.pvalue}{digits for P value, default 3.}

\item{digits.effect}{digits for effect value (e.g., OR, HR, or RR), default 2.}

\item{ref.value}{reference value.}

\item{...}{arguments passed to method from \code{\link[=lm]{lm()}}, \code{\link[=glm]{glm()}} or \code{\link[survival:coxph]{survival::coxph()}}.}
}
\value{
a data frame.
}
\description{
Multivariable analysis for coxph
}
\keyword{internal}
