% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{poson}
\alias{poson}
\title{Fitting poisson models}
\usage{
poson(data, formula, positive = "auto", ...)
}
\arguments{
\item{data}{a data frame, list or environment (or object coercible by
as.data.frame to a data frame) containing the variables in the model.}

\item{formula}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted.}

\item{positive}{in which positive of outcome variable to make the comparison.
By default, positive is automatically defined. If outcome is a factor variable,
then positive is defined as the highest level. If outcome is a numerical
variable, then positive is defined as the largest value.}

\item{...}{additional arguments to be passed to \code{\link[=glm]{glm()}} function.}
}
\description{
Fitting poisson models
}
\examples{
library(srmisc)

data("cancer")

cancer$sex  <- factor(cancer$sex)
cancer$race <- factor(cancer$race)
cancer$meta <- factor(cancer$meta)

# Fit a modified Poissson model
fit <- poson(cancer, meta ~ age + sex + race + size)

# Univariable modified Poissson regression
univariable(fit)

# Multivariable modified Poissson regression
multivariable(fit)
}
\seealso{
[glm[]]
}
