% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmatch.R
\name{print.psmatch}
\alias{print.psmatch}
\title{Print psmatch object}
\usage{
\method{print}{psmatch}(x, ...)
}
\arguments{
\item{x}{a psmatch object.}

\item{...}{further argumets pass to print function.}
}
\description{
Print psmatch object
}
