% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmatch.R
\name{psm}
\alias{psm}
\title{Matching for causal inference}
\usage{
psm(
  data,
  outcome = NULL,
  time = NULL,
  exposure = NULL,
  covariates = NULL,
  positive = "auto",
  method = "nearest",
  distance = "glm",
  link = "logit",
  caliper = 0.2,
  std.caliper = TRUE,
  ratio = 1,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{outcome}{outcome variable name.}

\item{time}{time variable name, for Cox regression.}

\item{exposure}{exposure variable name.}

\item{covariates}{covariate names, a vector or a list.}

\item{positive}{in which positive of outcome variable to make the comparison.
By default, positive is automatically defined. If outcome is a factor variable,
then positive is defined as the highest level. If outcome is a numerical
variable, then positive is defined as the largest value.}

\item{method}{the matching method to be used. Detail see \code{\link[MatchIt:matchit]{MatchIt::matchit()}}.}

\item{distance}{the distance measure to be used. \code{\link[MatchIt:matchit]{MatchIt::matchit()}}.}

\item{link}{when distance is specified as a string, an additional argument
controlling the link function used in estimating the distance measure.}

\item{caliper}{for methods that allow it, the width(s) of the caliper(s) to
use in matching. Should be a numeric vector with each value named according
to the variable to which the caliper applies. To apply to the distance measure,
the value should be unnamed. See the individual methods pages for information
on whether and how this argument is used. The default is NULL for no caliper.}

\item{std.caliper}{logical; when a caliper is specified, whether the the
caliper is in standard deviation units (TRUE) or raw units (FALSE).
Can either be of length 1, applying to all calipers, or of length equal
to the length of caliper. Default is TRUE.}

\item{ratio}{for methods that allow it, how many control units should be
matched to each treated unit in k:1 matching. Should be a single integer
value. See the individual methods pages for information on whether and how
this argument is used. The default is 1 for 1:1 matching.}

\item{...}{additional arguments passed to the \code{\link[MatchIt:matchit]{MatchIt::matchit()}}.}
}
\value{
a list.
}
\description{
Matching for causal inference
}
