% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm_baseline.R
\name{psm_baseline}
\alias{psm_baseline}
\title{Baseline characteritics for matched data}
\usage{
psm_baseline(
  object,
  psm = TRUE,
  iptw = TRUE,
  digits.smd = 3,
  digits.numeric = 2,
  digits.category = 1,
  ...
)
}
\arguments{
\item{object}{a psmatch object; the output of a call to \code{\link[=psm]{psm()}}.}

\item{psm}{logical, show baseline characteritics after PS.}

\item{iptw}{logical, show baseline characteritics after IPTW.}

\item{digits.smd}{digits for SMD, default 3.}

\item{digits.numeric}{digits for numeric variable, default 2.}

\item{digits.category}{digits for category variable, default 1.}

\item{...}{arguments pass to \code{\link[srr2bline:baseline]{srr2bline::baseline()}}}
}
\value{
a data frame.
}
\description{
Baseline characteritics for matched data
}
