% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm_data.R
\name{psm_data}
\alias{psm_data}
\title{Construct a matched dataset from a psmatch object}
\usage{
psm_data(
  object,
  type = c("psm", "iptw", "unmatched"),
  drop.unmatched = TRUE,
  ...
)
}
\arguments{
\item{object}{a psmatch object; the output of a call to \code{\link[=matched]{matched()}}.}

\item{type}{type matched dataset.}

\item{drop.unmatched}{logical; whether the returned data frame should contain
all units (FALSE) or only units that were matched (i.e., have a matching
weight greater than zero) (TRUE). Default is TRUE to drop unmatched units.}

\item{...}{unused.}
}
\value{
a data frame.
}
\description{
Construct a matched dataset from a psmatch object
}
