% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm-estimate.R
\name{psm_estimate}
\alias{psm_estimate}
\title{Effects for matched data}
\usage{
psm_estimate(object, model = "auto", ...)
}
\arguments{
\item{object}{a psmatch object; the output of a call to \code{\link[=psm]{psm()}}.}

\item{model}{model regression.}

\item{...}{further armuments pass to \code{\link[=typeset]{typeset()}}.}
}
\value{
a data frame.
}
\description{
Effects for matched data
}
