% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{typeset}
\alias{write_docx}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{srmisc}{\code{\link[srmisc]{typeset}}, \code{\link[srmisc]{write_docx}}}
}}

