% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskreg.R
\name{taskreg}
\alias{taskreg}
\title{Task for regression}
\usage{
taskreg(
  data,
  outcome = NULL,
  time = NULL,
  exposure = NULL,
  covariates = NULL,
  strata = NULL,
  positive = "auto",
  model = c("auto", "linear", "logit", "cox", "poson", "logbinom", "multinom"),
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{outcome}{outcome variable name.}

\item{time}{time variable name, for Cox regression.}

\item{exposure}{exposure variable name.}

\item{covariates}{covariate names, a vector or a list.}

\item{strata}{stratified varaible names.}

\item{positive}{in which positive of outcome variable to make the comparison.
By default, positive is automatically defined. If outcome is a factor variable,
then positive is defined as the highest level. If outcome is a numerical
variable, then positive is defined as the largest value.}

\item{model}{model regression.}

\item{...}{unused.}
}
\value{
a list
}
\description{
Task for regression
}
