% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIF.R
\name{vif}
\alias{vif}
\title{Variance inflation factors}
\usage{
vif(fit, ...)
}
\arguments{
\item{fit}{a regression model.}

\item{...}{not used.}
}
\value{
a data frame.
}
\description{
Calculates variance-inflation and generalized variance-inflation factors
(VIFs and GVIFs) for linear, generalized linear, and other regression models.
}
