% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate.R
\name{associate}
\alias{associate}
\title{Association analysis}
\usage{
associate(
  data,
  outcome = NULL,
  time = NULL,
  exposure = NULL,
  covariates = NULL,
  args = list(),
  p.trend = TRUE,
  n.quantile = NULL,
  quantile.right = TRUE,
  quantile.labels = NULL,
  conf.level = 0.95,
  conf.brackets = NULL,
  conf.separator = NULL,
  digits.pvalue = 3,
  digits.effect = 2,
  ref.value = 1,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{outcome}{outcome variable name.}

\item{time}{time variable name, for Cox regression.}

\item{exposure}{exposure variable name.}

\item{covariates}{covariate names, a vector or a list.}

\item{args}{arguments passed to method from \code{\link[=lm]{lm()}}, \code{\link[=glm]{glm()}} or \code{\link[survival:coxph]{survival::coxph()}}.}

\item{p.trend}{Tests for linear trend when exposure is numeric variable,
and set n.quantile.}

\item{n.quantile}{indicates how many quantiles to convert when exposure is
numeric variable.}

\item{quantile.right}{logical, indicating if the intervals should be closed
on the right (and open on the left) or vice versa.}

\item{quantile.labels}{labels for the levels of the resulting category.
By default, labels are constructed using "(a,b]" interval notation. If labels
= FALSE, simple integer codes are returned instead of a factor.}

\item{conf.level}{The confidence level to use for the confidence interval if
conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults to
0.95, which corresponds to a 95 percent confidence interval.}

\item{conf.brackets}{brackets of CI format.}

\item{conf.separator}{separate of CI format.}

\item{digits.pvalue}{digits for P value, default 3.}

\item{digits.effect}{digits for effect value (e.g., OR, HR, or RR), default 2.}

\item{ref.value}{reference value.}

\item{...}{further arguments.}
}
\value{
a data.frame.
}
\description{
Calculate the association between the two variables by
regression models (e.g. linear regression, logistic regression, Cox regression).
}
\examples{
library(srmisc)

cancer$sex  <- factor(cancer$sex)
cancer$race <- factor(cancer$race)
cancer$meta <- factor(cancer$meta)

# Association Between Metastasis and Tumor size by Linear Regression Model
# Model 1 adjusted for age, sex.
# Model 2 adjusted for age, sex, and race.
associate(data = cancer,
          outcome = "size",
          exposure = "meta",
          covariates = list(c("sex", "age"), c("sex", "age", "race")))

# Association Between Race and Metastasis by Logistic Regression Model
# Model 1 adjusted for nothing.
# Model 2 adjusted for age, sex, and size.
associate(data = cancer,
          outcome = "meta",
          exposure = "race",
          covariates = list("Adjusted 1" = NULL,
                            "Adjusted 2" = c("sex", "age", "size")))

# Association Between Race and Survival status by Cox Proportional Hazards Regression Model
# Model 1 adjusted for nothing.
# Model 2 adjusted for age, sex, and size.
associate(data = cancer,
          outcome = "status",
          time = "time",
          exposure = "race",
          covariates = list("Adjusted 1" = NULL,
                            "Adjusted 2" = c("sex", "age", "size")))

# Tests for linear trend were done by modeling the median value of each quantile
# to test ordered relations across quantiles of size.
associate(data = cancer,
          outcome = "status",
          time = "time",
          exposure = "size",
          covariates = list("Adjusted 1" = NULL,
                            "Adjusted 2" = c("sex", "age", "race")),
          n.quantile = 4)
}
