% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbinom.R
\name{logbinom}
\alias{logbinom}
\title{Fitting log-binomial models}
\usage{
logbinom(data, formula, ...)
}
\arguments{
\item{data}{a data frame, list or environment (or object coercible by
as.data.frame to a data frame) containing the variables in the model.}

\item{formula}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted.}

\item{...}{additional arguments to be passed to \code{\link[=glm]{glm()}} function.}
}
\description{
Fitting log-binomial models
}
