% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Fitting logistic models}
\usage{
logit(data, formula, ...)
}
\arguments{
\item{data}{a data frame, list or environment (or object coercible by
as.data.frame to a data frame) containing the variables in the model.}

\item{formula}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted.}

\item{...}{additional arguments to be passed to \code{\link[=glm]{glm()}} function.}
}
\value{
an object of class inheriting from "glm" which inherits from the
class "lm".
}
\description{
Fitting logistic models
}
